insta360-studio 5.7.0
This updates `insta360-studio` to the latest version, 5.7.0.

The `livecheck` block was broken partly because upstream switched
from a pkg file to a zip but also because the newest channel object
doesn't provide a `version` value. This updates the `livecheck` block
regex to match zip files and capture the version from the filename,
using that information in the `strategy` block in place of the
previous `version` value.

Besides that, the app is unsigned, so this adds a `disable!` call
with the future date we've been using for this situation.